Split string encryption provider into encryption and decryption providers
