+++ Application/EdFi.Ods.Api/Providers/ISymmetricStringEncryptionProvider.cs
@@ -20,13 +20,4 @@ public interface ISymmetricStringEncryptionProvider
     /// <returns>A string representing the input in encrypted form along with any
     /// other information necessary to decrypt it (aside from the private key)</returns>
     public string Encrypt(string value, byte[] key);
-    
-    /// <summary>
-    /// Decrypts the specified string value using the specified key.
-    /// </summary>
-    /// <para name="value">The string value to be decrypted.</para>
-    /// <para name="output">TThe decrypted string value</para>
-    /// <para name="key">The private key to be used for decryption.</para>
-    /// <returns>Indicates if the decryption operation was successful</returns>
-    public bool TryDecrypt(string value, out string output, byte[] key);
 }
